+++ CHANGELOG.md
@@ -4,8 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.17.74] - 2021-11-18
+## [0.17.75] - 2021-11-18
 - Fix font-size on footer
+
+## [0.17.74] - 2021-11-18
+- Fix video player
+- Add git player
+- Fix spelling issue in pdf report
 
 ## [0.17.73] - 2021-11-17
 - Modified divier line in PDF report